convert PageHeader to TS; add showTopMargin prop
