+++ src/components/PageHeader/PageHeader.scss
@@ -5,12 +5,6 @@
 @use '~@quartz/styles/scss/tokens';
 
 .container {
-	margin-top: 20px;
-
-	@include media-queries.tablet-landscape-up {
-		margin-top: 48px;
-	}
-
 	&.border {
 		border-bottom: borders.$solid-decorative;
 	}
@@ -20,6 +14,14 @@
 
 		@include media-queries.tablet-landscape-up {
 			padding-bottom: 60px;
+		}
+	}
+
+	&.show-top-margin {
+		margin-top: 20px;
+
+		@include media-queries.tablet-landscape-up {
+			margin-top: 48px;
 		}
 	}
 }
