+++ src/components/PageHeader/PageHeader.stories.mdx
@@ -1,5 +1,5 @@
 import { Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
-import PageHeader from './PageHeader.jsx';
+import PageHeader from './PageHeader';
 import { TabNav, TabNavItem } from '../TabNav/TabNav';
 
 <Meta title="Layout/PageHeader" component={PageHeader} />
